Refactored files structure. Divided Movie Details components into separated files
