+++ package.json
@@ -21,10 +21,14 @@
     "react-dom": "18.2.0",
     "react-icons": "^4.8.0",
     "react-query": "^3.39.3",
+    "react-slick": "^0.29.0",
     "sass": "^1.58.3",
+    "slick-carousel": "^1.8.1",
+    "swiper": "^9.1.1",
     "typescript": "4.9.5"
   },
   "devDependencies": {
+    "@types/react-slick": "^0.23.10",
     "autoprefixer": "^10.4.14",
     "postcss": "^8.4.21",
     "tailwindcss": "^3.2.7"
